+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
-
+### Added
+- `From` impls to create a `Literal` from references to proc-macro literal types:
+    - `From<&proc_macro::Literal>`
+    - `From<&proc_macro2::Literal>`
+- Better examples in README and repository
 
 ## 0.1.0 - 2021-05-24
 ### Added